+++ src/radarpub.rs
@@ -707,7 +707,7 @@ async fn tf_static(
     msg: ZBytes,
     enc: Encoding,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let topic = "rt/tf_static".to_string();
+    let topic = "tf_static".to_string();
     let mut interval = tokio::time::interval(Duration::from_secs(1));
 
     loop {
@@ -724,7 +724,7 @@ async fn radar_info(
     msg: ZBytes,
     enc: Encoding,
 ) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
-    let topic = "rt/radar/info".to_string();
+    let topic = "radar/info".to_string();
     let mut interval = tokio::time::interval(Duration::from_secs(1));
 
     loop {